+++ src/texstudio.h
@@ -316,6 +316,7 @@ protected slots:
 	void fileOpenAllRecent(); ///< open all files in recent file list
 	void fileRecentList();
     void fileClearRecentList();
+	void fileRemoveFromRecentList(const QString &fn);
 	void viewDocumentListHidden(); ///< show names of all hidden document (for debug)
 	void fileDocumentOpenFromChoosen(const QString &doc, int duplicate, int lineNr, int column);
 	void viewDocumentList();